All of Extended Task
First I fixed some of the naming from my previous commit that I forgot to replace in FizzBuzzValuesForRange Testing. Then I added everything for the Extended Task. I set up FizzBuzzifyAnInt to work with the dictionary, and use a constructor to set the default dictionary, and overloaded the constructor to set the dictionary manually. I then wrote the 3 unit test cases.